requirements.txt
google-analytics-data==0.18.14
google-auth==2.29.0
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
PyYAML==6.0.2
